fix(protocol): 修复AAD字段字节序问题并验证非负性
- 在AeadAad.encode()方法中添加ByteOrder.BIG_ENDIAN参数确保字节序正确
- 在AeadAad.decode()方法中添加负值检查，防止AAD字段为负数时引发异常
- 从AESUtil类中移除未使用的Logger导入
- 重构generateKey()方法支持算法参数化，提高代码复用性

docs(api): 更新API文档中的JSON字段命名规范

- 将所有camelCase字段名改为snake_case格式
- 更新sessionId为session_id，startTime为start_time，joinTime为join_time等
- 修改API文档注释明确JSON字段使用snake_case命名约定
- 更新api-chat-message.md中的时间戳字段命名一致性

chore(docs): 移除重复的客户端API文档文件

- 删除doc/clients/api-client.md文件，内容已整合到主API文档中

+++ connection/src/main/java/team/carrypigeon/backend/connection/protocol/codec/NettyEncoder.java
@@ -8,7 +8,7 @@ public class NettyEncoder extends MessageToByteEncoder<byte[]> {
 
     @Override
     protected void encode(ChannelHandlerContext channelHandlerContext, byte[] bytes, ByteBuf byteBuf) {
-        byteBuf.writeShort((short)bytes.length);
+        byteBuf.writeShort(bytes.length);
         byteBuf.writeBytes(bytes);
     }
 }
